requirement.txt
bcrypt==4.1.2
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
httpserver==1.1.0
idna==3.6
mysql-connector-python==8.2.0
numpy==1.26.3
opencv-python==4.9.0.80
protobuf==4.21.12
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.1.0
